+++ src/cli/account.rs
@@ -5,7 +5,7 @@ use crypto::{
     utils::{bytes_to_hex_string, Serializable},
     Felt,
 };
-use miden_client::{AuthInfo, Client};
+use miden_client::{client::Client, store::AuthInfo};
 use miden_lib::{faucets, AuthScheme};
 use objects::{
     accounts::{AccountId, AccountType},
