+++ src/client/transactions.rs
@@ -1,0 +1,56 @@
+use objects::{accounts::AccountId, notes::Note, transaction::TransactionScript, Digest};
+
+use super::Client;
+use crate::errors::ClientError;
+
+pub struct TransactionStub {
+    pub id: Digest,
+    pub account_id: AccountId,
+    pub init_account_state: Digest,
+    pub final_account_state: Digest,
+    pub input_note_nullifiers: Vec<Digest>,
+    pub output_notes: Vec<Note>,
+    pub transaction_script: Option<TransactionScript>,
+    pub block_num: u32,
+    pub committed: bool,
+    pub commit_height: u64,
+}
+
+impl TransactionStub {
+    #[allow(clippy::too_many_arguments)]
+    pub fn new(
+        id: Digest,
+        account_id: AccountId,
+        init_account_state: Digest,
+        final_account_state: Digest,
+        input_note_nullifiers: Vec<Digest>,
+        output_notes: Vec<Note>,
+        transaction_script: Option<TransactionScript>,
+        block_num: u32,
+        committed: bool,
+        commit_height: u64,
+    ) -> TransactionStub {
+        TransactionStub {
+            id,
+            account_id,
+            init_account_state,
+            final_account_state,
+            input_note_nullifiers,
+            output_notes,
+            transaction_script,
+            block_num,
+            committed,
+            commit_height,
+        }
+    }
+}
+
+impl Client {
+    // TRANSACTION DATA RETRIEVAL
+    // --------------------------------------------------------------------------------------------
+
+    /// Returns input notes managed by this client.
+    pub fn get_transactions(&self) -> Result<Vec<TransactionStub>, ClientError> {
+        self.store.get_transactions().map_err(|err| err.into())
+    }
+}
