+++ src/mock.rs
@@ -7,7 +7,10 @@ use miden_node_proto::{
 
 use objects::{utils::collections::BTreeMap, StarkField};
 
-use crate::{Client, FILTER_ID_SHIFT};
+use crate::{
+    client::{Client, FILTER_ID_SHIFT},
+    store::AuthInfo,
+};
 
 /// Mock RPC API
 ///
@@ -116,6 +119,6 @@ pub fn insert_mock_data(client: &mut Client) {
         .map_err(|err| format!("Error generating KeyPair: {}", err))
         .unwrap();
     client
-        .insert_account(&account, &crate::AuthInfo::RpoFalcon512(key_pair))
+        .insert_account(&account, &AuthInfo::RpoFalcon512(key_pair))
         .unwrap();
 }
